Moved trans() function to lang.go
